feat: upgrade to 0.3.2

+++ contracts/Interfaces/Compound/CEtherI.sol
@@ -10,5 +10,5 @@ interface CEtherI is CTokenI {
 
     function liquidateBorrow(address borrower, CTokenI cTokenCollateral) external payable;
 
-    function mint() external payable returns (uint);
+    function mint() external payable returns (uint256);
 }
